--- src/AC/articles.js
@@ -1,8 +1,0 @@
-import { DELETE_ARTICLE } from '../constants'
-
-export function deleteArticle(id) {
-    return {
-        type: DELETE_ARTICLE,
-        payload: { id }
-    }
-}
