+++ src/containers/Counter.js
@@ -1,31 +1,46 @@
-import React, { Component, PropTypes } from 'react'
-import { connect } from 'react-redux'
-import { increment } from '../AC/counter'
+import React, { Component } from "react";
+import PropTypes from "prop-types";
+import { connect } from "react-redux";
+import { increment, decrement } from "../actions";
 
 class Counter extends Component {
-    static propTypes = {
-        count: PropTypes.number,
-        increment: PropTypes.func
-    };
+  static propTypes = {
+    count: PropTypes.number,
+    increment: PropTypes.func
+  };
 
-    render() {
-        return (
-            <div>
-                <h1>{this.props.count}</h1>
-                <a href="#" onClick = {this.handleIncrement}>increment</a>
-            </div>
-        )
-    }
+  handleIncrement = e => {
+    e.preventDefault();
+    this.props.increment();
+  };
 
-    handleIncrement = (ev) => {
-        ev.preventDefault()
-        this.props.increment()
-    }
+  handleDecrement = e => {
+    e.preventDefault();
+    this.props.decrement();
+  };
+
+  render() {
+    return (
+      <div>
+        <h1>{this.props.count}</h1>
+        <a href="#" onClick={this.handleIncrement}>
+          increment
+        </a>
+        <a href="#" onClick={this.handleDecrement}>
+          decrement
+        </a>
+      </div>
+    );
+  }
 }
 
-export default connect((state) => {
-    const { count } = state
-    return { count }
-}, {
-    increment
-})(Counter)
+const mapStateToProps = state => {
+  return {
+    count: state.counter.count
+  };
+};
+
+export default connect(
+  mapStateToProps,
+  { increment, decrement }
+)(Counter);
