+++ src/containers/Root.js
@@ -1,24 +1,15 @@
-import React, { Component, PropTypes } from 'react'
-import { Provider } from 'react-redux'
-import Counter from './Counter'
-import Articles from './Articles'
-import store from '../store'
+import React from "react";
+import Articles from "./Articles";
+import Counter from "./Counter";
+import { Provider } from "react-redux";
+import store from "../store";
 
-class RootContainer extends Component {
-    static propTypes = {
-
-    };
-
-    render() {
-        return (
-            <Provider store = {store}>
-                <div>
-                    <Counter />
-                    <Articles />
-                </div>
-            </Provider>
-        )
-    }
+function Root() {
+  return (
+    <Provider store={store}>
+      <Articles />
+    </Provider>
+  );
 }
 
-export default RootContainer
+export default Root;
